backend/requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.1
aniso8601==0.92
itsdangerous==0.24
python-firebase==1.2
pytz==2014.10
requests==2.5.1
six==1.9.0
wsgiref==0.1.2
yolk==0.4.3
